+++ src/confluent_kafka/schema_registry/schema_registry_client.py
@@ -15,7 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
+import abc
 import json
 import logging
 import random
@@ -30,12 +30,14 @@ from collections import defaultdict
 from enum import Enum
 from threading import Lock
 from typing import List, Dict, Type, TypeVar, \
-    cast, Optional, Union, Any, Tuple
+    cast, Optional, Union, Any, Tuple, Callable
 
 from cachetools import TTLCache, LRUCache
 from httpx import Response
 
-from .error import SchemaRegistryError
+from authlib.integrations.httpx_client import OAuth2Client
+
+from .error import SchemaRegistryError, OAuthTokenError
 
 # TODO: consider adding `six` dependency or employing a compat file
 # Python 2.7 is officially EOL so compatibility issue will be come more the norm.
@@ -58,6 +60,72 @@ except NameError:
 
 log = logging.getLogger(__name__)
 VALID_AUTH_PROVIDERS = ['URL', 'USER_INFO']
+
+
+class _BearerFieldProvider(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def get_bearer_fields(self) -> dict:
+        raise NotImplementedError
+
+
+class _StaticFieldProvider(_BearerFieldProvider):
+    def __init__(self, token: str, logical_cluster: str, identity_pool: str):
+        self.token = token
+        self.logical_cluster = logical_cluster
+        self.identity_pool = identity_pool
+
+    def get_bearer_fields(self) -> dict:
+        return {'bearer.auth.token': self.token, 'bearer.auth.logical.cluster': self.logical_cluster,
+                'bearer.auth.identity.pool.id': self.identity_pool}
+
+
+class _CustomOAuthClient(_BearerFieldProvider):
+    def __init__(self, custom_function: Callable[[Dict], Dict], custom_config: dict):
+        self.custom_function = custom_function
+        self.custom_config = custom_config
+
+    def get_bearer_fields(self) -> dict:
+        return self.custom_function(self.custom_config)
+
+
+class _OAuthClient(_BearerFieldProvider):
+    def __init__(self, client_id: str, client_secret: str, scope: str, token_endpoint: str, logical_cluster: str,
+                 identity_pool: str, max_retries: int, retries_wait_ms: int, retries_max_wait_ms: int):
+        self.token = None
+        self.logical_cluster = logical_cluster
+        self.identity_pool = identity_pool
+        self.client = OAuth2Client(client_id=client_id, client_secret=client_secret, scope=scope)
+        self.token_endpoint = token_endpoint
+        self.max_retries = max_retries
+        self.retries_wait_ms = retries_wait_ms
+        self.retries_max_wait_ms = retries_max_wait_ms
+        self.token_expiry_threshold = 0.8
+
+    def get_bearer_fields(self) -> dict:
+        return {'bearer.auth.token': self.get_access_token(), 'bearer.auth.logical.cluster': self.logical_cluster,
+                'bearer.auth.identity.pool.id': self.identity_pool}
+
+    def token_expired(self) -> bool:
+        expiry_window = self.token['expires_in'] * self.token_expiry_threshold
+
+        return self.token['expires_at'] < time.time() + expiry_window
+
+    def get_access_token(self) -> str:
+        if not self.token or self.token_expired():
+            self.generate_access_token()
+
+        return self.token['access_token']
+
+    def generate_access_token(self) -> None:
+        for i in range(self.max_retries + 1):
+            try:
+                self.token = self.client.fetch_token(url=self.token_endpoint, grant_type='client_credentials')
+                return
+            except Exception as e:
+                if i >= self.max_retries:
+                    raise OAuthTokenError(f"Failed to retrieve token after {self.max_retries} "
+                                          f"attempts due to error: {str(e)}")
+                time.sleep(full_jitter(self.retries_wait_ms, self.retries_max_wait_ms, i) / 1000)
 
 
 class _BaseRestClient(object):
@@ -150,7 +218,7 @@ class _BaseRestClient(object):
         self.max_retries = 3
         max_retries = conf_copy.pop('max.retries', None)
         if max_retries is not None:
-            if not isinstance(timeout, (int, float)):
+            if not isinstance(max_retries, (int, float)):
                 raise TypeError("max.retries must be a number, not " + str(type(max_retries)))
             self.max_retries = max_retries
 
@@ -169,6 +237,86 @@ class _BaseRestClient(object):
                 raise TypeError("retries.max.wait.ms must be a number, not "
                                 + str(type(retries_max_wait_ms)))
             self.retries_max_wait_ms = retries_max_wait_ms
+
+        self.bearer_field_provider = None
+        logical_cluster = None
+        identity_pool = None
+        self.bearer_auth_credentials_source = conf_copy.pop('bearer.auth.credentials.source', None)
+        if self.bearer_auth_credentials_source is not None:
+            self.auth = None
+
+            if self.bearer_auth_credentials_source in {'OAUTHBEARER', 'STATIC_TOKEN'}:
+                headers = ['bearer.auth.logical.cluster', 'bearer.auth.identity.pool.id']
+                missing_headers = [header for header in headers if header not in conf_copy]
+                if missing_headers:
+                    raise ValueError("Missing required bearer configuration properties: {}"
+                                     .format(", ".join(missing_headers)))
+
+                logical_cluster = conf_copy.pop('bearer.auth.logical.cluster')
+                if not isinstance(logical_cluster, str):
+                    raise TypeError("logical cluster must be a str, not " + str(type(logical_cluster)))
+
+                identity_pool = conf_copy.pop('bearer.auth.identity.pool.id')
+                if not isinstance(identity_pool, str):
+                    raise TypeError("identity pool id must be a str, not " + str(type(identity_pool)))
+
+            if self.bearer_auth_credentials_source == 'OAUTHBEARER':
+                properties_list = ['bearer.auth.client.id', 'bearer.auth.client.secret', 'bearer.auth.scope',
+                                   'bearer.auth.issuer.endpoint.url']
+                missing_properties = [prop for prop in properties_list if prop not in conf_copy]
+                if missing_properties:
+                    raise ValueError("Missing required OAuth configuration properties: {}".
+                                     format(", ".join(missing_properties)))
+
+                self.client_id = conf_copy.pop('bearer.auth.client.id')
+                if not isinstance(self.client_id, string_type):
+                    raise TypeError("bearer.auth.client.id must be a str, not " + str(type(self.client_id)))
+
+                self.client_secret = conf_copy.pop('bearer.auth.client.secret')
+                if not isinstance(self.client_secret, string_type):
+                    raise TypeError("bearer.auth.client.secret must be a str, not " + str(type(self.client_secret)))
+
+                self.scope = conf_copy.pop('bearer.auth.scope')
+                if not isinstance(self.scope, string_type):
+                    raise TypeError("bearer.auth.scope must be a str, not " + str(type(self.scope)))
+
+                self.token_endpoint = conf_copy.pop('bearer.auth.issuer.endpoint.url')
+                if not isinstance(self.token_endpoint, string_type):
+                    raise TypeError("bearer.auth.issuer.endpoint.url must be a str, not "
+                                    + str(type(self.token_endpoint)))
+
+                self.bearer_field_provider = _OAuthClient(self.client_id, self.client_secret, self.scope,
+                                                          self.token_endpoint, logical_cluster, identity_pool,
+                                                          self.max_retries, self.retries_wait_ms,
+                                                          self.retries_max_wait_ms)
+            elif self.bearer_auth_credentials_source == 'STATIC_TOKEN':
+                if 'bearer.auth.token' not in conf_copy:
+                    raise ValueError("Missing bearer.auth.token")
+                static_token = conf_copy.pop('bearer.auth.token')
+                self.bearer_field_provider = _StaticFieldProvider(static_token, logical_cluster, identity_pool)
+                if not isinstance(static_token, string_type):
+                    raise TypeError("bearer.auth.token must be a str, not " + str(type(static_token)))
+            elif self.bearer_auth_credentials_source == 'CUSTOM':
+                custom_bearer_properties = ['bearer.auth.custom.provider.function',
+                                            'bearer.auth.custom.provider.config']
+                missing_custom_properties = [prop for prop in custom_bearer_properties if prop not in conf_copy]
+                if missing_custom_properties:
+                    raise ValueError("Missing required custom OAuth configuration properties: {}".
+                                     format(", ".join(missing_custom_properties)))
+
+                custom_function = conf_copy.pop('bearer.auth.custom.provider.function')
+                if not callable(custom_function):
+                    raise TypeError("bearer.auth.custom.provider.function must be a callable, not "
+                                    + str(type(custom_function)))
+
+                custom_config = conf_copy.pop('bearer.auth.custom.provider.config')
+                if not isinstance(custom_config, dict):
+                    raise TypeError("bearer.auth.custom.provider.config must be a dict, not "
+                                    + str(type(custom_config)))
+
+                self.bearer_field_provider = _CustomOAuthClient(custom_function, custom_config)
+            else:
+                raise ValueError('Unrecognized bearer.auth.credentials.source')
 
         # Any leftover keys are unknown to _RestClient
         if len(conf_copy) > 0:
@@ -208,6 +356,23 @@ class _RestClient(_BaseRestClient):
             proxy=self.proxy,
             timeout=self.timeout
         )
+
+    def handle_bearer_auth(self, headers: dict) -> None:
+        bearer_fields = self.bearer_field_provider.get_bearer_fields()
+        required_fields = ['bearer.auth.token', 'bearer.auth.identity.pool.id', 'bearer.auth.logical.cluster']
+
+        missing_fields = []
+        for field in required_fields:
+            if field not in bearer_fields:
+                missing_fields.append(field)
+
+        if missing_fields:
+            raise ValueError("Missing required bearer auth fields, needs to be set in config or custom function: {}"
+                             .format(", ".join(missing_fields)))
+
+        headers["Authorization"] = "Bearer {}".format(bearer_fields['bearer.auth.token'])
+        headers['Confluent-Identity-Pool-Id'] = bearer_fields['bearer.auth.identity.pool.id']
+        headers['target-sr-cluster'] = bearer_fields['bearer.auth.logical.cluster']
 
     def get(self, url: str, query: Optional[dict] = None) -> Any:
         return self.send_request(url, method='GET', query=query)
@@ -255,6 +420,9 @@ class _RestClient(_BaseRestClient):
             body = json.dumps(body)
             headers = {'Content-Length': str(len(body)),
                        'Content-Type': "application/vnd.schemaregistry.v1+json"}
+
+        if self.bearer_auth_credentials_source:
+            self.handle_bearer_auth(headers)
 
         response = None
         for i, base_url in enumerate(self.base_urls):
@@ -316,7 +484,7 @@ class _RestClient(_BaseRestClient):
         for i in range(self.max_retries + 1):
             response = self.session.request(
                 method, url="/".join([base_url, url]),
-                headers=headers, data=body, params=query)
+                headers=headers, content=body, params=query)
 
             if is_success(response.status_code):
                 return response
@@ -373,7 +541,7 @@ class _SchemaCache(object):
             self.schema_id_index[subject][schema_id] = schema
             self.schema_index[subject][schema] = schema_id
 
-    def set_registered_schema(self, registered_schema: 'RegisteredSchema'):
+    def set_registered_schema(self, schema: 'Schema', registered_schema: 'RegisteredSchema'):
         """
         Add a RegisteredSchema to the cache.
 
@@ -384,7 +552,6 @@ class _SchemaCache(object):
         subject = registered_schema.subject
         schema_id = registered_schema.schema_id
         version = registered_schema.version
-        schema = registered_schema.schema
         with self.lock:
             self.schema_id_index[subject][schema_id] = schema
             self.schema_index[subject][schema] = schema_id
@@ -686,7 +853,8 @@ class SchemaRegistryClient(object):
 
         registered_schema = RegisteredSchema.from_dict(response)
 
-        self._cache.set_schema(subject_name, registered_schema.schema_id, registered_schema.schema)
+        # The registered schema may not be fully populated
+        self._cache.set_schema(subject_name, registered_schema.schema_id, schema)
 
         return registered_schema
 
@@ -764,9 +932,17 @@ class SchemaRegistryClient(object):
                                           .format(_urlencode(subject_name), normalize_schemas, deleted),
                                           body=request)
 
-        registered_schema = RegisteredSchema.from_dict(response)
+        result = RegisteredSchema.from_dict(response)
 
-        self._cache.set_registered_schema(registered_schema)
+        # Ensure the schema matches the input
+        registered_schema = RegisteredSchema(
+            schema_id=result.schema_id,
+            subject=result.subject,
+            version=result.version,
+            schema=schema,
+        )
+
+        self._cache.set_registered_schema(schema, registered_schema)
 
         return registered_schema
 
@@ -925,7 +1101,7 @@ class SchemaRegistryClient(object):
 
         registered_schema = RegisteredSchema.from_dict(response)
 
-        self._cache.set_registered_schema(registered_schema)
+        self._cache.set_registered_schema(registered_schema.schema, registered_schema)
 
         return registered_schema